+++ src/mind2pu_function.ps1
@@ -49,18 +49,6 @@
     If you replace '*' with '-', it becomes a map that extends from
     right to left. This is equivalent to specifying -RightToLeft switch.
 
-    Legend can be output by writing the following:
-
-        legend right|left
-        this is legend
-        end legend
-    
-    Fill color can be specified by prepending or postponing
-    [#color] to the label. For example:
-
-        - [#orange] label
-        - label [#blue]
-
 .LINK
     pu2java, dot2gviz, pert, pert2dot, pert2gantt2pu, mind2dot, mind2pu, gantt2pu, logi2dot, logi2dot2, logi2dot3, logi2pu, logi2pu2, flow2pu
 
@@ -77,7 +65,6 @@
 .PARAMETER FoldLabel
     Fold label at specified number of characters.
 
-
 .PARAMETER Kinsoku
     Wrapping of character string considering japanese KINSOKU rules.
     Specify numerical value as 1 for ASCII characters
@@ -85,12 +72,6 @@
 
     Depends on kinsoku_function.ps1
 
-
-.PARAMETER LegendRight
-    Insert legend in bottom right
-
-.PARAMETER LegendLeft
-    Insert legend in bottom left
 
 .PARAMETER RightToLeft
     Right to left graph
@@ -119,7 +100,7 @@
     @startmindmap
 
     title "title"
-    skinparam DefaultFontName "Meiryo"
+    skinparam DefaultFontName "MS Gothic"
 
     * hogehoge
     ** hoge1
@@ -185,7 +166,7 @@ function mind2pu {
         [string]$FontName,
 
         [Parameter( Mandatory=$False)]
-        [string]$FontNameWindowsDefault = "Meiryo",
+        [string]$FontNameWindowsDefault = "MS Gothic",
 
         [Parameter( Mandatory=$False)]
         [int]$FontSize,
@@ -210,12 +191,6 @@ function mind2pu {
 
         [Parameter( Mandatory=$False)]
         [int]$KinsokuOnlyPlainText,
-
-        [Parameter( Mandatory=$False)]
-        [string[]]$LegendRight,
-
-        [Parameter( Mandatory=$False)]
-        [string[]]$LegendLeft,
 
         [Parameter( Mandatory=$False)]
         [switch]$RightToLeft,
@@ -337,12 +312,24 @@ function mind2pu {
             ## set node
             $readLineAryNode += $ast + $contents
         } else {
+            ## is legend block?
+            if (($isLegend) -and ($rdLine -match '^end *legend$')){
+                $readLineAryLeg += $rdLine
+                $isLegend = $false
+                return
+            }
+            if ($isLegend){
+                $readLineAryLeg += $rdLine
+                return
+            }
+            if (($lineCounter -gt 1) -and ($rdLine -match '^legend (right|left)$')){
+                $readLineAryLeg = @()
+                $readLineAryLeg += $rdLine
+                $isLegend = $True
+                return
+            }
             ## output as is
             $readLineAryNode += $rdLine
-            ## is legend block?
-            if (($lineCounter -gt 1) -and ($rdLine -match '^legend (right|left)$')){
-                $isLegend = $True
-            }
         }
     }
     end {
@@ -396,28 +383,6 @@ function mind2pu {
         ##
         ## Footer
         ##
-        if($LegendLeft){
-            $legendFlag = $True
-            $readLineAryLegend = @()
-            $readLineAryLegend += ""
-            $readLineAryLegend += "legend left"
-            foreach($legLine in $LegendLeft){
-                $readLineAryLegend += $legLine
-            }
-            $readLineAryLegend += "end legend"
-        }elseif($LegendRight){
-            $legendFlag = $True
-            $readLineAryLegend = @()
-            $readLineAryLegend += ""
-            $readLineAryLegend += "legend right"
-            foreach($legLine in $LegendRight){
-                $readLineAryLegend += $legLine
-            }
-            $readLineAryLegend += "end legend"
-        }else{
-            $legendFlag = $False
-        }
-
         $readLineAryFooter = @()
         $readLineAryFooter += ""
         if ($WBS) {
@@ -432,9 +397,9 @@ function mind2pu {
         foreach ($lin in $readLineAryNode){
             $readLineAry += $lin
         }
-        if($legendFlag){
-            foreach ($lin in $readLineAryLegend){
-                $readLineAry += $lin
+        if ( $readLineAryLeg.Count -gt 0 ){
+            foreach ( $leg in $readLineAryLeg ){
+                $readLineAry += $leg
             }
         }
         foreach ($lin in $readLineAryFooter){